Move cli's concurrent checker to its own file
